requirements.txt
requests_cache==1.0.0
tqdm==4.67.1
beautifulsoup4==4.12.3
lxml==5.3.0
lxml-html-clean==0.4.1
prettytable==3.14.0
sqlalchemy==1.4.54
requests-html==0.10.0
